improved kernel tests
